test: fix ExpectedCurrentVersion type in release tests
Use explicit uint64(0) when passing ExpectedCurrentVersion to
ActivateReleaseRequest so the value matches the expected protobuf
field type.

+++ internal/server/grpcserver/releases_test.go
@@ -72,7 +72,7 @@ func TestConfigurationReleaseGRPCLifecycleAndWatch(t *testing.T) {
 	if created.GetRelease().GetVersion() != 1 || created.GetRelease().GetDigest() == "" {
 		t.Fatalf("created=%+v", created.GetRelease())
 	}
-	active, err := releases.ActivateRelease(adminCtx(), &kmsv1.ActivateReleaseRequest{Namespace: pNS("prod", "app"), Name: "runtime", Version: 1, ExpectedCurrentVersion: new(0)})
+	active, err := releases.ActivateRelease(adminCtx(), &kmsv1.ActivateReleaseRequest{Namespace: pNS("prod", "app"), Name: "runtime", Version: 1, ExpectedCurrentVersion: new(uint64(0))})
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -83,7 +83,7 @@ func TestConfigurationReleaseGRPCLifecycleAndWatch(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
-	if _, err := releases.ActivateRelease(adminCtx(), &kmsv1.ActivateReleaseRequest{Namespace: pNS("prod", "app"), Name: "runtime", Version: createdV2.GetRelease().GetVersion(), ExpectedCurrentVersion: new(0)}); status.Code(err) != codes.Aborted {
+	if _, err := releases.ActivateRelease(adminCtx(), &kmsv1.ActivateReleaseRequest{Namespace: pNS("prod", "app"), Name: "runtime", Version: createdV2.GetRelease().GetVersion(), ExpectedCurrentVersion: new(uint64(0))}); status.Code(err) != codes.Aborted {
 		t.Fatalf("stale CAS code=%s err=%v, want Aborted", status.Code(err), err)
 	}
 	stream, err := releases.WatchRelease(adminCtx())
